Don't cast pgoff_t to possibly 32-bit types for output
pgoff_t is most likely a 64-bit integer, so casting it to a 32-bit
type for output could lose data.  In the cases addressed here, the
files cannot actually get that large, so this is only cosmetic and to
set better examples for the future.  (Similar issues that could have
actual practical impact were addressed separately in commit
e8f851d6172.)

In one case, the 32-bit size is baked into the protocol, so here we
add an elog and document this discrepancy.

Discussion: https://www.postgresql.org/message-id/flat/20ce62fa-47fc-457b-b504-12f3c1651726%40eisentraut.org

+++ src/backend/access/heap/rewriteheap.c
@@ -1104,8 +1104,8 @@ heap_xlog_logical_rewrite(XLogReaderState *r)
 	if (ftruncate(fd, xlrec->offset) != 0)
 		ereport(ERROR,
 				(errcode_for_file_access(),
-				 errmsg("could not truncate file \"%s\" to %u: %m",
-						path, (uint32) xlrec->offset)));
+				 errmsg("could not truncate file \"%s\" to %lld: %m",
+						path, (long long int) xlrec->offset)));
 	pgstat_report_wait_end();
 
 	data = XLogRecGetData(r) + sizeof(*xlrec);
